rename origin Home to home
